pyproject.toml
[project]
name = "bottec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "apscheduler>=3.11.0",
    "django>=5.2.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.9",
    "python-decouple>=3.8",
    "python-slugify>=8.0.4",
    "uvicorn>=0.34.2",
]
